requirements.txt
autopep8==1.6.0
paho-mqtt==1.6.1
pycodestyle==2.8.0
toml==0.10.2
wakeonlan==2.1.0
